refactor: template 파일에서 dom 컨트롤하는 함수를 view로 이동
